Separated the llri headers
