+++ legion/engine/llri/detail/adapter_extensions.hpp
