+++ legion/engine/llri/detail/device.inl
@@ -1,0 +1,7 @@
+#pragma once
+#include <llri/llri.hpp> //Recursive include technically not necessary but helps with intellisense
+
+namespace legion::graphics::llri
+{
+
+}
