+++ legion/engine/llri/detail/instance.inl
@@ -1,46 +1,8 @@
 #pragma once
+#include <llri/llri.hpp> //Recursive include technically not necessary but helps with intellisense
 
 namespace legion::graphics::llri
 {
-    constexpr inline const char* to_string(const result& result)
-    {
-        switch (result)
-        {
-        case result::Success:
-            return "Success";
-        case result::Timeout:
-            return "Timeout";
-        case result::ErrorUnknown:
-            return "ErrorUnknown";
-        case result::ErrorInvalidUsage:
-            return "ErrorInvalidUsage";
-        case result::ErrorFeatureNotSupported:
-            return "ErrorFeatureNotSupported";
-        case result::ErrorExtensionNotSupported:
-            return "ErrorExtensionNotSupported";
-        case result::ErrorDeviceHung:
-            return "ErrorDeviceHung";
-        case result::ErrorDeviceLost:
-            return "ErrorDeviceLost";
-        case result::ErrorDeviceRemoved:
-            return "ErrorDeviceRemoved";
-        case result::ErrorDriverFailure:
-            return "ErrorDriverFailure";
-        case result::NotReady:
-            return "NotReady";
-        case result::ErrorOutOfHostMemory:
-            return "ErrorOutOfHostMemory";
-        case result::ErrorOutOfDeviceMemory:
-            return "ErrorOutOfDeviceMemory";
-        case result::ErrorInitializationFailed:
-            return "ErrorInitializationFailed";
-        case result::ErrorIncompatibleDriver:
-            return "ErrorIncompatibleDriver";
-        }
-
-        return "Invalid result value";
-    }
-
     constexpr inline const char* to_string(const validation_callback_severity& severity)
     {
         switch (severity)
@@ -71,23 +33,6 @@ namespace legion::graphics::llri
         }
 
         return "Invalid validation_callback_source value";
-    }
-
-    constexpr inline const char* to_string(const adapter_type& type)
-    {
-        switch (type)
-        {
-        case adapter_type::Other:
-            return "Other";
-        case adapter_type::Integrated:
-            return "Integrated";
-        case adapter_type::Discrete:
-            return "Discrete";
-        case adapter_type::Virtual:
-            return "Virtual";
-        }
-
-        return "Invalid adapter_type value";
     }
 
     constexpr inline const char* to_string(const instance_extension_type& result)
@@ -126,7 +71,7 @@ namespace legion::graphics::llri
 #endif
 
 #ifndef LLRI_DISABLE_INTERNAL_API_MESSAGE_POLLING
-        const auto r =  detail::impl_createInstance(desc, instance, true);
+        const auto r = detail::impl_createInstance(desc, instance, true);
         detail::impl_pollAPIMessages((*instance)->m_validationCallback, (*instance)->m_validationCallbackMessenger);
         return r;
 #else
@@ -203,81 +148,6 @@ namespace legion::graphics::llri
 #ifndef LLRI_DISABLE_INTERNAL_API_MESSAGE_POLLING
         //Can't use device messenger here because the device is destroyed
         detail::impl_pollAPIMessages(m_validationCallback, m_validationCallbackMessenger);
-#endif
-    }
-
-    inline result Adapter::queryInfo(adapter_info* info) const
-    {
-#ifndef LLRI_DISABLE_VALIDATION
-        if (info == nullptr)
-        {
-            m_validationCallback(validation_callback_severity::Error, validation_callback_source::Validation, "Adapter::queryInfo() returned ErrorInvalidUsage because the passed info parameter was nullptr.");
-            return result::ErrorInvalidUsage;
-        }
-
-        if (m_ptr == nullptr)
-        {
-            m_validationCallback(validation_callback_severity::Error, validation_callback_source::Validation, "Adapter::queryInfo() returned ErrorDeviceLost because the passed adapter has a nullptr internal handle which usually indicates a lost device.");
-            return result::ErrorDeviceLost;
-        }
-#endif
-
-#ifndef LLRI_DISABLE_INTERNAL_API_MESSAGE_POLLING
-        const auto r = impl_queryInfo(info);
-        detail::impl_pollAPIMessages(m_validationCallback, m_validationCallbackMessenger);
-        return r;
-#else
-        return impl_queryInfo(info);
-#endif
-    }
-
-    inline result Adapter::queryFeatures(adapter_features* features) const
-    {
-#ifndef LLRI_DISABLE_VALIDATION
-        if (features == nullptr)
-        {
-            m_validationCallback(validation_callback_severity::Error, validation_callback_source::Validation, "Adapter::queryFeatures() returned ErrorInvalidUsage because the passed features parameter was nullptr.");
-            return result::ErrorInvalidUsage;
-        }
-
-        if (m_ptr == nullptr)
-        {
-            m_validationCallback(validation_callback_severity::Error, validation_callback_source::Validation, "Adapter::queryFeatures() returned ErrorDeviceLost because the passed adapter has a nullptr internal handle which usually indicates a lost device.");
-            return result::ErrorDeviceLost;
-        }
-#endif
-
-#ifndef LLRI_DISABLE_INTERNAL_API_MESSAGE_POLLING
-        const auto r = impl_queryFeatures(features);
-        detail::impl_pollAPIMessages(m_validationCallback, m_validationCallbackMessenger);
-        return r;
-#else
-        return impl_queryFeatures(features);
-#endif
-    }
-
-    inline result Adapter::queryExtensionSupport(const adapter_extension_type& type, bool* supported) const
-    {
-#ifndef LLRI_DISABLE_VALIDATION
-        if (supported == nullptr)
-        {
-            m_validationCallback(validation_callback_severity::Error, validation_callback_source::Validation, "Adapter::queryExtensionSupport() returned ErrorInvalidUsage because the passed supported parameter was nullptr.");
-            return result::ErrorInvalidUsage;
-        }
-
-        if (m_ptr == nullptr)
-        {
-            m_validationCallback(validation_callback_severity::Error, validation_callback_source::Validation, "Adapter::queryExtensionSupport() returned ErrorDeviceLost because the passed adapter has a nullptr internal handle which usually indicates a lost device.");
-            return result::ErrorDeviceLost;
-        }
-#endif
-
-#ifndef LLRI_DISABLE_INTERNAL_API_MESSAGE_POLLING
-        const auto r = impl_queryExtensionSupport(type, supported);
-        detail::impl_pollAPIMessages(m_validationCallback, m_validationCallbackMessenger);
-        return r;
-#else
-        return impl_queryExtensionSupport(type, supported);
 #endif
     }
 }
