+++ legion/engine/llri/llri.vcxproj
@@ -122,10 +122,16 @@ del /q ".\xcopyexclude"</Command>
     </PreBuildEvent>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClInclude Include="adapter_extensions.hpp" />
-    <ClInclude Include="instance_extensions.hpp" />
+    <ClInclude Include="detail\adapter.hpp" />
+    <ClInclude Include="detail\adapter_extensions.hpp" />
+    <ClInclude Include="detail\adapter.inl" />
+    <ClInclude Include="detail\device.hpp" />
+    <ClInclude Include="detail\device.inl" />
+    <ClInclude Include="detail\instance.hpp" />
+    <ClInclude Include="detail\instance_extensions.hpp" />
+    <ClInclude Include="detail\instance.inl" />
     <ClInclude Include="llri.hpp" />
-    <ClInclude Include="llri_impl.hpp" />
+    <ClInclude Include="detail\llri.inl" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
